==> preprocessing_pipeline/6_split_workplaces.ipynb <==
### Here we take dataframe and split workplaces into smaller compartments

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
from shapely.geometry import Point
from tqdm import tqdm
from geopy.distance import great_circle
import time as time
import random
from scipy.stats import binom

city_name = 'chelyabinsk'

input_path_people = '../data/initial_' + city_name + '/people_' + city_name + '_assigned_schools.txt'
input_path_workplaces = '../data/initial_' + city_name + '/workplaces_' + city_name + '.txt'

people = pd.read_csv(input_path_people, sep='\t', index_col=0)
workplaces = pd.read_csv(input_path_workplaces, sep='\t')


people['school_id'] = people['school_id'].str.replace('X','-1')
people['school_id'] = people['school_id'].astype(str).astype(float).astype(int)

people['work_id'] = people['work_id'].str.replace('X','-1')
people['work_id'] = people['work_id'].astype(str).astype(float).astype(int)

print(people.head())
print("Total number of people in population:", len(people['sp_id'].unique()))
print(workplaces.head())
print("Total number of workplaces in population", len(workplaces))
people['school_id'].unique()

def generate_distribution_data(sum, num_of_samples):
    data = []
    n_i = num_of_samples
    y_i = sum
    for i in range(num_of_samples):
        x_i = binom.rvs(y_i, 1/n_i, size=1)[0]
        data.append(x_i)
        y_i = y_i - x_i
        n_i = n_i - 1
    return data

work_id_old = people['work_id'].unique()
work_id_old = work_id_old
work_id_old = [int(float(elem)) for elem in work_id_old]
people_splitted = pd.DataFrame(columns=['sp_id', 'sp_hh_id', 'age',	'sex',	'race',	'relate', 'school_id', 'work_id'])
workplaces_splitted = pd.DataFrame(columns=['sp_id', 'latitude', 'longitude', 'size'])
work_id_max = 1

people_list = []

work_id_list = []
work_latitude_list = []
work_longitude_list = []
work_size_list = []

people_whole = people.copy()
people = people_whole[people['work_id'] > 0]
people_without_workplace = people_whole[people_whole['work_id'] < 0]
print(len(people_without_workplace), len(people))

for work_id in tqdm(work_id_old):
    if (work_id == -1):
        people_in_current_workplace = people[people['work_id'] == work_id]
        print(len(people_in_current_workplace))
        people_list.extend(list(people_in_current_workplace['sp_id']))
        for i in range(len(people_in_current_workplace)):
            work_id_list.append(-1)
    else:   
        people_in_current_workplace = people[people['work_id'] == work_id]
        people_list.extend(list(people_in_current_workplace['sp_id']))
        # print(list(people_in_current_dwelling['sp_id']))
        number_of_people_in_workplace = len(people_in_current_workplace)
        # print(people_list)
        
        
        if (number_of_people_in_workplace < 10):
            distribution_for_new_work_id = [number_of_people_in_workplace]
        else:
            distribution_for_new_work_id = generate_distribution_data(number_of_people_in_workplace, int(number_of_people_in_workplace/8.5))
        
        # distribution_for_new_hh_id = np.random.poisson(2.5, number_of_people_in_dwelling) 
        # print(number_of_people_in_workplace, sum(distribution_for_new_work_id))
        if (number_of_people_in_workplace != sum(distribution_for_new_work_id)):
            raise RuntimeError("Mismatch")
        
        for people_in_work_id in distribution_for_new_work_id:
            for i in range(people_in_work_id):
                work_id_list.append(work_id_max)
                work_size_list.append(people_in_work_id)
                work_latitude_list.append(float(workplaces[workplaces['sp_id'] == work_id].latitude))
                work_longitude_list.append(float(workplaces[workplaces['sp_id'] == work_id].longitude))
            work_id_max+=1

people_splitted = people
people_splitted['work_id'] = work_id_list
workplaces_splitted['sp_id'] = work_id_list
workplaces_splitted['size'] = work_size_list
workplaces_splitted['latitude'] = work_latitude_list
workplaces_splitted['longitude'] = work_longitude_list

people_splitted = pd.concat([people_splitted, people_without_workplace])

people_splitted.to_csv('../data/initial_' + city_name + '/people_' + city_name + '_splitted_workplaces.txt', sep='\t')
workplaces_splitted.to_csv('../data/initial_' + city_name + '/workplaces_' + city_name + '_splitted.txt', sep='\t')

people_splitted['school_id'].unique()
