==> preprocessing_pipeline/7_final_preparation.ipynb <==
### Here we make final preparation before simulation of epidemics in the city

import numpy as np
import pandas as pd 
import matplotlib.pyplot as plt
from shapely.geometry import Point
from tqdm import tqdm
from geopy.distance import great_circle
import time as time
import random
from scipy.stats import binom

city_name = 'chelyabinsk'

input_path = '../data/initial_' + city_name + '/people_' + city_name + '_splitted_workplaces.txt'

people = pd.read_csv(input_path, sep='\t', index_col=0)

def move_school_id_to_work_id(row):
    value = 0
    if row['school_id'] != -1:
        value = row['school_id']
    else:
        value = row['work_id']
    return value


people['work_id'] = people.apply(move_school_id_to_work_id, axis  = 1)
people_processed = people.drop(columns=['Unnamed: 0', 'race', 'relate'])
people_processed.head()

def change_work_to_X(row):
    value = 0
    if (row['work_id'] == -1):
        value = 'X'
    else:
        value = row['work_id']
    return value

def change_school_to_X(row):
    value = 0
    if (row['school_id'] == -1):
        value = 'X'
    else:
        value = row['school_id']
    return value
    
    
people_processed['work_id'] = people_processed.apply(change_work_to_X, axis  = 1)
people_processed['school_id'] = people_processed.apply(change_school_to_X, axis  = 1)
people_processed.sort_index(inplace=True)

# len(people_processed[people_processed['work_id'] == 'X'])

##### And also we need to create schools.json file for model with school number as key and values are sp_id's

school_dict = dict()


for idx in tqdm(range(len(people_processed))):
    row = people_processed.iloc[idx]
    school_id = str(row['school_id'])
    if (school_id != 'X' and school_id != -1):
        if (school_id in school_dict):
            school_dict[school_id].append(int(idx))
        else:
            school_dict[school_id] = [int(idx)]

# del school_dict['X']

people_processed

import json

with open('../data/initial_' + city_name + '/schools.json', 'w') as fp:
    json.dump(school_dict, fp)
    
people_processed = people_processed.drop(columns=['school_id'])
people_processed.head(15)
people_processed.to_csv('../data/initial_' + city_name + '/people.txt', sep='\t')
